minor bug fixes due to the refactor
